Bug Patch (Pressure Sensor)
Fixed "conflicting return type"

+++ Main/Peripherals/include/Drivers/ExPressureSensor.h
@@ -8,7 +8,7 @@
 class ExPressureSensor : public IPressureSensor {
 public:
     ExPressureSensor();
-    float getPressure();
+    int getPressure();
 private:
 };
